Refactor services architecture: replace base `CrudService` with `EntityRepository` pattern
