+++ src/lib/domain/clan-management/repository/ClanRepository.ts
@@ -1,0 +1,5 @@
+import {EntityRepository} from "@/lib/domain/clan-management/repository/EntityRepository";
+import {Clan} from "@/types";
+
+export interface ClanRepository extends EntityRepository<Clan> {
+}
