+++ src/lib/domain/clan-management/repository/ClanarinaRepository.ts
@@ -1,0 +1,5 @@
+import {EntityRepository} from "@/lib/domain/clan-management/repository/EntityRepository";
+import {Clanarina} from "@/types";
+
+export interface ClanarinaRepository extends EntityRepository<Clanarina> {
+}
