+++ src/lib/domain/clan-management/repository/EntityRepository.ts
@@ -1,0 +1,9 @@
+export interface EntityRepository<T> {
+    save(entity: T): Promise<T>;
+
+    delete(id: string): Promise<void>;
+
+    find(id: string): Promise<T | null>;
+
+    findAll(): Promise<T[]>;
+}
