+++ src/lib/persistence/clanarina/google-spreadsheet/ClanarinaSpreadsheetConfig.ts
@@ -1,0 +1,13 @@
+import {SpreadsheetConfig} from "@/lib/persistence/google-spreadsheet/SpreadsheetConfig";
+import {Clanarina} from "@/types";
+import {ClanarinaDataTransformer} from "@/lib/persistence/clanarina/google-spreadsheet/ClanarinaDataTransformer";
+
+export const ClanarinaSpreadsheetConfig: SpreadsheetConfig<Clanarina> = {
+    sheetConfig: {
+        sheetName: 'Clanarine',
+        range: 'Clanarine!A:C',
+        dateColumn: {index: 2, letter: 'C'}
+    },
+    dataTransformer: new ClanarinaDataTransformer(),
+    idField: 'id'
+}
