+++ src/lib/persistence/google-spreadsheet/DataTransformer.ts
@@ -1,0 +1,4 @@
+export interface DataTransformer<TEntity> {
+    rowToEntity: (row: string[]) => TEntity;
+    entityToRow: (entity: TEntity) => string[];
+}
