--- src/lib/services/index.ts
@@ -1,5 +1,0 @@
-import { ClanarineService } from './ClanarineService';
-import { ClanoviService } from './ClanoviService';
-
-export const clanarine = new ClanarineService(process.env.GOOGLE_SPREADSHEET_ID!)
-export const clanovi = new ClanoviService(process.env.GOOGLE_SPREADSHEET_ID!)
